+++ Accounting-Service/src/main/java/de/grado/accountingservice/repository/JournalEntryLineRepository.java
@@ -4,9 +4,7 @@ import de.grado.accountingservice.model.JournalEntryLine;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigInteger;
-
 @Repository
-public interface JournalEntryLineRepository extends JpaRepository<JournalEntryLine, BigInteger>
+public interface JournalEntryLineRepository extends JpaRepository<JournalEntryLine, Long>
 {
 }
