+++ Accounting-Service/src/main/java/de/grado/accountingservice/repository/JournalEntryRepository.java
@@ -4,9 +4,7 @@ import de.grado.accountingservice.model.JournalEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.math.BigDecimal;
-
 @Repository
-public interface JournalEntryRepository extends JpaRepository<JournalEntry, BigDecimal>
+public interface JournalEntryRepository extends JpaRepository<JournalEntry, Long>
 {
 }
